requirements.txt
# Telegram Bot
python-telegram-bot==20.7

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0

# Google Sheets API
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Scheduler
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.0
pytz==2023.3

# Telegram Session Automation
telethon==1.34.0

# Development (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
